requirements.txt
Django==2.0.6
dj-database-url==0.5.0
django-cors-headers==2.4.0
djangorestframework==3.8.2
psycopg2==2.7.5
pytz==2018.5
